qwen-odoo-training/requirements.txt
# ============================================================
# Qwen 32B Odoo Fine-Tuning - Dependencies
# ============================================================
# Install: pip install -r requirements.txt
# Note: Install PyTorch with CUDA first from https://pytorch.org
# ============================================================

# --- Core ML ---
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# --- Hugging Face ecosystem ---
transformers>=4.44.0
datasets>=2.19.0
accelerate>=0.33.0
peft>=0.12.0
trl>=0.9.0
tokenizers>=0.19.0
safetensors>=0.4.0

# --- DeepSpeed for multi-GPU ---
deepspeed>=0.14.0

# --- Quantization ---
bitsandbytes>=0.43.0
auto-gptq>=0.7.0

# --- Flash Attention (build from source on Windows if needed) ---
# pip install flash-attn --no-build-isolation
# flash-attn>=2.5.0

# --- Monitoring & Logging ---
tensorboard>=2.16.0
wandb>=0.17.0
rich>=13.7.0
tqdm>=4.66.0

# --- Data processing ---
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.1.0
gitpython>=3.1.0
markdown>=3.6
html2text>=2024.2.26

# --- Evaluation ---
rouge-score>=0.1.2
nltk>=3.8.0
sacrebleu>=2.4.0
evaluate>=0.4.0
human-eval>=1.0

# --- Utilities ---
pyyaml>=6.0.0
psutil>=5.9.0
pynvml>=11.5.0
colorama>=0.4.6
tabulate>=0.9.0
